Shrink the size of deoptimization_counter_ and kind_tag_ fields. This compensates for the new num_optional_named_parameters_ field. Review URL: https://chromiumcodereview.appspot.com//10905220
git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@12226 260f80e4-7a28-3924-810f-c04153c831b5

+++ runtime/vm/object.h
@@ -1500,10 +1500,10 @@ class Function : public Object {
     raw_ptr()->usage_counter_ = value;
   }
 
-  intptr_t deoptimization_counter() const {
+  int16_t deoptimization_counter() const {
     return raw_ptr()->deoptimization_counter_;
   }
-  void set_deoptimization_counter(intptr_t value) const {
+  void set_deoptimization_counter(int16_t value) const {
     raw_ptr()->deoptimization_counter_ = value;
   }
