+++ runtime/vm/raw_object.h
@@ -616,8 +616,8 @@ class RawFunction : public RawObject {
   intptr_t num_optional_positional_parameters_;
   intptr_t num_optional_named_parameters_;
   intptr_t usage_counter_;  // Incremented while function is running.
-  intptr_t deoptimization_counter_;
-  intptr_t kind_tag_;
+  uint16_t deoptimization_counter_;
+  uint16_t kind_tag_;
 };
